requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.4

# LLM — OpenRouter uses the OpenAI-compatible API
openai==1.51.0

# Document Extraction
pdfplumber==0.10.4
PyMuPDF==1.23.26
python-docx==1.1.0

# OCR (optional AWS)
boto3==1.34.51
pytesseract==0.3.10
Pillow==10.4.0

# NLP & Embeddings — fastembed uses ONNX runtime (no torch, no CUDA, ~50MB)
fastembed==0.3.6

# Validation & Parsing
pydantic==2.6.4
pydantic-settings==2.2.1
phonenumbers==8.13.31
dateparser==1.2.0
chardet==5.2.0

# Async Queue
arq==0.25.0
redis==5.0.2

# Utilities
tenacity==8.2.3
loguru==0.7.2
python-dotenv==1.0.1

# Testing
pytest==8.1.0
pytest-asyncio==0.23.5
httpx==0.27.0
